@@ -5,7 +5,6 @@ import 'package:await_route/await_route.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
-import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:refresh_storage/refresh_storage.dart';
 import 'package:utils/utils.dart';
 
@@ -28,8 +27,8 @@ class _ChanceStorage extends RefreshStorageItem {
 class NativeAdBuilder extends StatefulObserverWidget {
   /// Creates [NativeAdBuilder].
   const NativeAdBuilder({
-    Key key,
-    @required this.builder,
+    Key? key,
+    required this.builder,
     this.controller,
     this.preloadCount = 0,
   }) : super(key: key);
@@ -44,24 +43,29 @@ class NativeAdBuilder extends StatefulObserverWidget {
   final NativeAdChildBuilder builder;
 
   /// [NativeAdController] of [NativeAdBuilder]. If not specified, constructs its own controller.
-  final NativeAdController controller;
+  final NativeAdController? controller;
 
   /// The amount of [NativeAdController] to keep preloaded, after this controller is used.
   final int preloadCount;
 
   /// Checks whether to show an ad based on chance.
-  static bool checkChance([double rate]) => random.nextDouble() < (rate ?? defaultRate);
+  static bool checkChance([double? rate]) => random.nextDouble() < (rate ?? defaultRate);
 
   /// Checks whether the [NativeAdController] for this identifier is already in the page storage.
-  static bool hasController(BuildContext context, String identifier, {RefreshStorageState storage}) =>
+  static bool hasController(BuildContext context, String identifier, {RefreshStorageState? storage}) =>
       RefreshStorage.contains(context, identifier, storage: storage);
 
   /// Checks whether the [NativeAdController] for this identifier is already in
   /// the page storage or the [checkChance] returns true for this call.
-  static bool checkChanceFor(BuildContext context, String identifier, [double rate]) =>
+  static bool checkChanceFor(BuildContext context, String identifier, [double? rate]) =>
       hasController(context, identifier) || checkChance(rate);
 
-  static bool checkChanceForPage(BuildContext context, String identifier, {double rate, RefreshStorageState storage}) {
+  static bool checkChanceForPage(
+    BuildContext context,
+    String identifier, {
+    double? rate,
+    RefreshStorageState? storage,
+  }) {
     // Check if this ad has already been made.
     if (hasController(context, identifier, storage: storage)) return true;
 
@@ -73,7 +77,7 @@ class NativeAdBuilder extends StatefulObserverWidget {
       storage: storage,
     );
 
-    final truthy = entry.value.truthy;
+    final truthy = entry.value!.truthy;
     entry.dispose();
     return truthy;
   }
@@ -104,8 +108,8 @@ class NativeAdBuilder extends StatefulObserverWidget {
   static Widget childBuilder(
     int i,
     int n, {
-    @required Widget Function(int index) adBuilder,
-    @required Widget Function(int index) childBuilder,
+    required Widget Function(int index) adBuilder,
+    required Widget Function(int index) childBuilder,
     int offset = 0,
     bool enabled = true,
   }) {
@@ -154,14 +158,14 @@ class NativeAdBuilder extends StatefulObserverWidget {
 }
 
 class _NativeAdBuilderState extends State<NativeAdBuilder> with InitialDependencies<NativeAdBuilder> {
-  NativeAdController _controller;
+  late final NativeAdController _controller;
   bool _hadAttachedToTheController = false;
 
-  Future _preloadControllers() => NativeAdBuilder.preloadControllers(widget.preloadCount, widget.controller.optionsKey);
-  bool _isControllerReady() => _controller.nativeAd?.maybeMap((_) => true, orElse: () => false);
+  Future _preloadControllers() => NativeAdBuilder.preloadControllers(widget.preloadCount, _controller.optionsKey);
+  bool _isControllerReady() => _controller.nativeAd.maybeMap((_) => true, orElse: () => false) ?? false;
 
   /// Skip fetching the ad, if the user is scrolling too quick.
-  Future _deferredLoad([Duration _]) async {
+  Future _deferredLoad([Duration? _]) async {
     if (!mounted) {
       return;
     } else {
@@ -169,7 +173,7 @@ class _NativeAdBuilderState extends State<NativeAdBuilder> with InitialDependenc
 
       final isReady = _isControllerReady(); // Check again.
       if (!isReady && Scrollable.recommendDeferredLoadingForContext(context)) {
-        WidgetsBinding.instance.addPostFrameCallback(_deferredLoad);
+        WidgetsBinding.instance!.addPostFrameCallback(_deferredLoad);
       } else {
         _controller.attach(this);
         _hadAttachedToTheController = true;
@@ -178,13 +182,13 @@ class _NativeAdBuilderState extends State<NativeAdBuilder> with InitialDependenc
     }
   }
 
-  Future _deferredIdleLoad([Duration _]) async {
+  Future _deferredIdleLoad([Duration? _]) async {
     if (!mounted) {
       return;
     } else {
       await AwaitRoute.of(context);
       if (Scrollable.recommendIdleLoadingForContext(context)) {
-        WidgetsBinding.instance.addPostFrameCallback(_deferredIdleLoad);
+        WidgetsBinding.instance!.addPostFrameCallback(_deferredIdleLoad);
       } else {
         await _preloadControllers();
       }
@@ -193,7 +197,7 @@ class _NativeAdBuilderState extends State<NativeAdBuilder> with InitialDependenc
 
   @override
   void initDependencies() {
-    _controller = widget.controller ?? NativeAdController.firstReusable() ?? NativeAdController();
+    _controller = widget.controller ?? NativeAdController.reuseOrCreate();
     _deferredLoad();
     if (widget.preloadCount > 0) _deferredIdleLoad();
   }
@@ -206,8 +210,8 @@ class _NativeAdBuilderState extends State<NativeAdBuilder> with InitialDependenc
 
   @override
   void dispose() {
-    if (_hadAttachedToTheController) _controller?.detach(this);
-    if (widget.controller == null) _controller?.dispose();
+    if (_hadAttachedToTheController) _controller.detach(this);
+    if (widget.controller == null) _controller.dispose();
     super.dispose();
   }
 
