@@ -10,9 +10,9 @@ import 'native_ad_builder.dart';
 /// [errorBuilder] & [loadingBuilder] can be left undefined to build an [SizedBox.shrink] instead.
 class NativeAdWidgetStateBuilder extends StatefulWidget implements NativeAdWidget {
   const NativeAdWidgetStateBuilder({
-    Key key,
-    @required this.identifier,
-    @required this.builder,
+    Key? key,
+    required this.identifier,
+    required this.builder,
     this.errorBuilder,
     this.loadingBuilder,
     this.options = NativeAdOptions.defaultKey,
@@ -26,13 +26,13 @@ class NativeAdWidgetStateBuilder extends StatefulWidget implements NativeAdWidge
   @override
   final String options;
   @override
-  final NativeAdController controller;
+  final NativeAdController? controller;
 
   final int preloadCount;
   final Widget Function(BuildContext, NativeAdController, NativeAdData) builder;
-  final Widget Function(BuildContext, NativeAdController, NativeAdErrorData) errorBuilder;
-  final Widget Function(BuildContext, NativeAdController, NativeAdLoadingData) loadingBuilder;
-  final Widget Function(BuildContext, Widget) layoutBuilder;
+  final Widget Function(BuildContext, NativeAdController, NativeAdErrorData)? errorBuilder;
+  final Widget Function(BuildContext, NativeAdController, NativeAdLoadingData)? loadingBuilder;
+  final Widget Function(BuildContext, Widget)? layoutBuilder;
 
   @override
   _NativeAdWidgetStateBuilderState createState() => _NativeAdWidgetStateBuilderState();
@@ -43,18 +43,18 @@ class _NativeAdWidgetStateBuilderState extends NativeAdWidgetState<NativeAdWidge
     final child = nativeAd.map(
       (data) => KeyedSubtree(
         key: ValueKey(data.headline), // FIXME: Find a more viable variable to assert equality with.
-        child: widget.builder(context, controller, data),
+        child: widget.builder(context, controller!, data),
       ),
       loading: (data) => widget.loadingBuilder != null
           ? KeyedSubtree(
               key: const ValueKey('loading'),
-              child: widget.loadingBuilder(context, controller, data),
+              child: widget.loadingBuilder!(context, controller!, data),
             )
           : const SizedBox.shrink(key: ValueKey('loading')),
       error: (data) => widget.errorBuilder != null
           ? KeyedSubtree(
               key: const ValueKey('error'),
-              child: widget.errorBuilder(context, controller, data),
+              child: widget.errorBuilder!(context, controller!, data),
             )
           : const SizedBox.shrink(key: ValueKey('error')),
     );
