@@ -8,8 +8,8 @@ part of 'native_ad.dart';
 
 _$_NativeAdImageDrawable _$_$_NativeAdImageDrawableFromJson(Map json) {
   return _$_NativeAdImageDrawable(
-    width: (json['width'] as num)?.toDouble(),
-    height: (json['height'] as num)?.toDouble(),
+    width: (json['width'] as num).toDouble(),
+    height: (json['height'] as num).toDouble(),
     bitmap: const Uint8ListConverter().fromJson(json['bitmap']),
   );
 }
@@ -25,12 +25,11 @@ Map<String, dynamic> _$_$_NativeAdImageDrawableToJson(
 _$_NativeAdImage _$_$_NativeAdImageFromJson(Map json) {
   return _$_NativeAdImage(
     uri: json['uri'] as String,
-    scale: (json['scale'] as num)?.toDouble(),
+    scale: (json['scale'] as num).toDouble(),
     drawable: json['drawable'] == null
         ? null
-        : NativeAdImageDrawable.fromJson((json['drawable'] as Map)?.map(
-            (k, e) => MapEntry(k as String, e),
-          )),
+        : NativeAdImageDrawable.fromJson(
+            Map<String, dynamic>.from(json['drawable'] as Map)),
   );
 }
 
@@ -43,10 +42,10 @@ Map<String, dynamic> _$_$_NativeAdImageToJson(_$_NativeAdImage instance) =>
 
 _$_NativeAdMediaContent _$_$_NativeAdMediaContentFromJson(Map json) {
   return _$_NativeAdMediaContent(
-    aspectRatio: (json['aspectRatio'] as num)?.toDouble(),
+    aspectRatio: (json['aspectRatio'] as num).toDouble(),
     hasVideoContent: json['hasVideoContent'] as bool,
-    duration: (json['duration'] as num)?.toDouble(),
-    mainImage: const Uint8ListConverter().fromJson(json['mainImage']),
+    duration: (json['duration'] as num).toDouble(),
+    mainImage: const Uint8ListConverterNullable().fromJson(json['mainImage']),
   );
 }
 
@@ -56,19 +55,17 @@ Map<String, dynamic> _$_$_NativeAdMediaContentToJson(
       'aspectRatio': instance.aspectRatio,
       'hasVideoContent': instance.hasVideoContent,
       'duration': instance.duration,
-      'mainImage': const Uint8ListConverter().toJson(instance.mainImage),
+      'mainImage':
+          const Uint8ListConverterNullable().toJson(instance.mainImage),
     };
 
 _$_AdChoicesInfo _$_$_AdChoicesInfoFromJson(Map json) {
   return _$_AdChoicesInfo(
-    text: json['text'] as String,
-    images: (json['images'] as List)
-            ?.map((e) => e == null
-                ? null
-                : NativeAdImage.fromJson((e as Map)?.map(
-                    (k, e) => MapEntry(k as String, e),
-                  )))
-            ?.toList() ??
+    text: json['text'] as String?,
+    images: (json['images'] as List<dynamic>?)
+            ?.map((e) =>
+                NativeAdImage.fromJson(Map<String, dynamic>.from(e as Map)))
+            .toList() ??
         [],
   );
 }
@@ -76,7 +73,7 @@ _$_AdChoicesInfo _$_$_AdChoicesInfoFromJson(Map json) {
 Map<String, dynamic> _$_$_AdChoicesInfoToJson(_$_AdChoicesInfo instance) =>
     <String, dynamic>{
       'text': instance.text,
-      'images': instance.images?.map((e) => e?.toJson())?.toList(),
+      'images': instance.images.map((e) => e.toJson()).toList(),
     };
 
 _$NativeAdData _$_$NativeAdDataFromJson(Map json) {
@@ -85,37 +82,31 @@ _$NativeAdData _$_$NativeAdDataFromJson(Map json) {
     body: json['body'] as String,
     icon: json['icon'] == null
         ? null
-        : NativeAdImage.fromJson((json['icon'] as Map)?.map(
-            (k, e) => MapEntry(k as String, e),
-          )),
-    callToAction: json['callToAction'] as String,
-    advertiser: json['advertiser'] as String,
-    price: json['price'] as String,
-    store: json['store'] as String,
-    starRating: (json['starRating'] as num)?.toDouble(),
-    isCustomClickGestureEnabled: json['isCustomClickGestureEnabled'] as bool,
-    isCustomMuteThisAdEnabled: json['isCustomMuteThisAdEnabled'] as bool,
+        : NativeAdImage.fromJson(
+            Map<String, dynamic>.from(json['icon'] as Map)),
+    callToAction: json['callToAction'] as String?,
+    advertiser: json['advertiser'] as String?,
+    price: json['price'] as String?,
+    store: json['store'] as String?,
+    starRating: (json['starRating'] as num?)?.toDouble(),
+    isCustomClickGestureEnabled: json['isCustomClickGestureEnabled'] as bool?,
+    isCustomMuteThisAdEnabled: json['isCustomMuteThisAdEnabled'] as bool?,
     adChoicesInfo: json['adChoicesInfo'] == null
         ? null
-        : AdChoicesInfo.fromJson((json['adChoicesInfo'] as Map)?.map(
-            (k, e) => MapEntry(k as String, e),
-          )),
+        : AdChoicesInfo.fromJson(
+            Map<String, dynamic>.from(json['adChoicesInfo'] as Map)),
     mediaContent: json['mediaContent'] == null
         ? null
-        : NativeAdMediaContent.fromJson((json['mediaContent'] as Map)?.map(
-            (k, e) => MapEntry(k as String, e),
-          )),
-    images: (json['images'] as List)
-            ?.map((e) => e == null
-                ? null
-                : NativeAdImage.fromJson((e as Map)?.map(
-                    (k, e) => MapEntry(k as String, e),
-                  )))
-            ?.toList() ??
+        : NativeAdMediaContent.fromJson(
+            Map<String, dynamic>.from(json['mediaContent'] as Map)),
+    images: (json['images'] as List<dynamic>?)
+            ?.map((e) =>
+                NativeAdImage.fromJson(Map<String, dynamic>.from(e as Map)))
+            .toList() ??
         [],
-    muteThisAdReasons: (json['muteThisAdReasons'] as List)
+    muteThisAdReasons: (json['muteThisAdReasons'] as List<dynamic>?)
             ?.map((e) => e as String)
-            ?.toList() ??
+            .toList() ??
         [],
   );
 }
@@ -134,7 +125,7 @@ Map<String, dynamic> _$_$NativeAdDataToJson(_$NativeAdData instance) =>
       'isCustomMuteThisAdEnabled': instance.isCustomMuteThisAdEnabled,
       'adChoicesInfo': instance.adChoicesInfo?.toJson(),
       'mediaContent': instance.mediaContent?.toJson(),
-      'images': instance.images?.map((e) => e?.toJson())?.toList(),
+      'images': instance.images.map((e) => e.toJson()).toList(),
       'muteThisAdReasons': instance.muteThisAdReasons,
     };
 
