@@ -5,7 +5,7 @@ import 'controller/controller.dart';
 
 class NativeAdWidgetStateStorage extends RefreshStorageItem {
   NativeAdWidgetStateStorage({String options = NativeAdOptions.defaultKey})
-      : controller = NativeAdController.reuseOrCreate(options: options ?? NativeAdOptions.defaultKey);
+      : controller = NativeAdController.reuseOrCreate(options: options);
 
   final NativeAdController controller;
 
@@ -19,15 +19,15 @@ class NativeAdWidgetStateStorage extends RefreshStorageItem {
 mixin NativeAdWidget on StatefulWidget {
   String get identifier;
   String get options => NativeAdOptions.defaultKey;
-  NativeAdController get controller;
+  NativeAdController? get controller;
 }
 
 /// State that persists a [NativeAdController] within the [RefreshStorage] as well as
 /// rebuilding it, when the controller is considered old on `initState` or app coming
 /// into the foreground.
 abstract class NativeAdWidgetState<T extends NativeAdWidget> extends State<T> {
-  RefreshStorageEntry<NativeAdWidgetStateStorage> storage;
-  NativeAdController get controller => widget.controller ?? storage?.value?.controller;
+  late final RefreshStorageEntry<NativeAdWidgetStateStorage> storage;
+  NativeAdController? get controller => widget.controller ?? storage.value?.controller;
 
   NativeAdWidgetStateStorage _buildStorage() => NativeAdWidgetStateStorage(options: widget.options);
 
@@ -40,7 +40,7 @@ abstract class NativeAdWidgetState<T extends NativeAdWidget> extends State<T> {
   void _checkOldController() {
     assert(widget.controller == null);
 
-    if (storage?.value?.controller?.considerThisOld() == true) {
+    if (storage.value?.controller.considerThisOld() == true) {
       storage.dispose();
       RefreshStorage.destroy(context: context, identifier: widget.identifier);
 
@@ -78,7 +78,7 @@ abstract class NativeAdWidgetState<T extends NativeAdWidget> extends State<T> {
   @mustCallSuper
   @override
   void dispose() {
-    storage?.dispose();
+    storage.dispose();
     super.dispose();
   }
 }
