@@ -9,14 +9,9 @@ part of 'platform_props.dart';
 _$_NativeAdPlatformProps _$_$_NativeAdPlatformPropsFromJson(Map json) {
   return _$_NativeAdPlatformProps(
     unitId: json['unitId'] as String,
-    options: (json['options'] as Map)?.map(
-      (k, e) => MapEntry(
-          k as String,
-          e == null
-              ? null
-              : NativeAdOptions.fromJson((e as Map)?.map(
-                  (k, e) => MapEntry(k as String, e),
-                ))),
+    options: (json['options'] as Map).map(
+      (k, e) => MapEntry(k as String,
+          NativeAdOptions.fromJson(Map<String, dynamic>.from(e as Map))),
     ),
   );
 }
@@ -25,20 +20,17 @@ Map<String, dynamic> _$_$_NativeAdPlatformPropsToJson(
         _$_NativeAdPlatformProps instance) =>
     <String, dynamic>{
       'unitId': instance.unitId,
-      'options': instance.options?.map((k, e) => MapEntry(k, e?.toJson())),
+      'options': instance.options.map((k, e) => MapEntry(k, e.toJson())),
     };
 
 _$_PlatformProps _$_$_PlatformPropsFromJson(Map json) {
   return _$_PlatformProps(
-    nativeAd: json['nativeAd'] == null
-        ? null
-        : NativeAdPlatformProps.fromJson((json['nativeAd'] as Map)?.map(
-            (k, e) => MapEntry(k as String, e),
-          )),
+    nativeAd: NativeAdPlatformProps.fromJson(
+        Map<String, dynamic>.from(json['nativeAd'] as Map)),
   );
 }
 
 Map<String, dynamic> _$_$_PlatformPropsToJson(_$_PlatformProps instance) =>
     <String, dynamic>{
-      'nativeAd': instance.nativeAd?.toJson(),
+      'nativeAd': instance.nativeAd.toJson(),
     };
